# small performance improvement