не создаем новый объект с аттрибутами

```
Warming up --------------------------------------
Light (associations): collection through .to_json
                         1.000  i/100ms
Blueprinter (associations): collection
                         1.000  i/100ms
Calculating -------------------------------------
Light (associations): collection through .to_json
                         12.617  (± 7.9%) i/s -     63.000  in   5.072740s
Blueprinter (associations): collection
                         12.399  (± 8.1%) i/s -     62.000  in   5.062292s

Comparison:
Light (associations): collection through .to_json:       12.6 i/s
Blueprinter (associations): collection:       12.4 i/s - same-ish: difference falls within error
```